finalizad health helpers
