+++ chronos_npm_package/controllers/healthHelpers.js
@@ -2,211 +2,201 @@ const si = require('systeminformation');
 
 const healthHelpers = {};
 
-/*
-This method returns an promise that resolves to an array of
-si data points.
-*/
+/* This object contains all systeminformation methods, metric names, and corresponding points */
+
+const collectedMetrics = {
+  cpu: {
+    speed_in_GHz: 'speed',
+    speedMax_in_GHz: 'speedMax',
+    num_of_cores: 'cores',
+    num_of_processors: 'processors',
+    'cache.l1d in bytes': 'cache.l1d',
+    'cache.l1i in bytes': 'cache.l1i',
+    'cache.l2 in bytes': 'cache.l2',
+    'cache.l3 in bytes': 'cache.l3',
+  },
+  cpuCurrentSpeed: {
+    average_CPU_speed_in_GHz: 'avg',
+    minimum_CPU_speed_in_GHz: 'min',
+    maximum_CPU_speed_in_GHz: 'max',
+  },
+  cpuTemperature: {
+    average_temperature: 'main',
+    max_temperature: 'max',
+  },
+  currentLoad: {
+    average_CPU_load_percent: 'avg',
+    current_CPU_load_percent: 'currentLoad',
+    current_CPU_load_user_percent: 'currentLoadUser',
+    current_CPU_load__system_percent: 'currentLoadSystem',
+    current_CPU_load_nice_percent: 'currentLoadNice',
+    current_CPU_load_idle_percent: 'currentLoadIdle',
+    current_CPU_load_raw_ticks: 'rawCurrentLoad',
+  },
+  mem: {
+    totalmemory_in_bytes: 'total',
+    freememory_in_bytes: 'free',
+    usedmemory_in_bytes: 'used',
+    activememory_in_bytes: 'active',
+    buffers_plus_cache_in_bytes: 'buffcache',
+    available_memory: 'available',
+  },
+  processes: {
+    totalprocesses: 'all',
+    blockedprocesses: 'blocked',
+    runningprocesses: 'running',
+    sleepingprocesses: 'sleeping',
+  },
+  inetLatency: 'all data collected',
+};
 
 healthHelpers.collectHealthData = () => {
-  // Create an array to hold the promises returned by si
-  const promises = [];
-  // for use with every object
+  const healthDataCollection = [];
   const time = Date.now();
-  // Fill the array of promises
-  promises.push(
-    si
-      .cpu()
-      .then(data => ({
-        speed_in_GHz: data.speed,
-        speedMax_in_GHz: data.speedMax,
-        num_of_cores: data.cores,
-        num_of_processors: data.processors,
-        'cache.l1d in bytes': data.cache.l1d,
-        'cache.l1i in bytes': data.cache.l1i,
-        'cache.l2 in bytes': data.cache.l2,
-        'cache.l3 in bytes': data.cache.l3,
-      }))
-      .then(data => {
-        const cpuMetrics = [];
-        for (const metric in data) {
-          cpuMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'CPU',
-            time,
-          });
-        }
-        return cpuMetrics;
-      })
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
-  promises.push(
-    si
-      .cpuCurrentSpeed()
-      .then(data => ({
-        average_CPU_speed_in_GHz: data.avg,
-        minimum_CPU_speed_in_GHz: data.min,
-        maximum_CPU_speed_in_GHz: data.max,
-      }))
-      .then(data => {
-        const cpuSpeedMetrics = [];
-        for (const metric in data) {
-          cpuSpeedMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'CPU',
-            time,
-          });
-        }
-        return cpuSpeedMetrics;
-      })
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
 
-  promises.push(
-    si
-      .cpuTemperature()
-      .then(data => ({
-        average_temperature: data.main,
-        max_temperature: data.max,
-      }))
-      .then(data => {
-        const cpuTemperatureMetrics = [];
-        for (const metric in data) {
-          cpuTemperatureMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'CPU',
-            time,
-          });
-        }
-        return cpuTemperatureMetrics;
-      })
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
+  console.log('in health helpers 2');
 
-  promises.push(
-    si
-      .currentLoad()
-      .then(data => ({
-        average_CPU_load_percent: data.avg,
-        current_CPU_load_percent: data.currentLoad,
-        current_CPU_load_user_percent: data.currentLoadUser,
-        current_CPU_load__system_percent: data.currentLoadSystem,
-        current_CPU_load_nice_percent: data.currentLoadNice,
-        current_CPU_load_idle_percent: data.currentLoadIdle,
-        current_CPU_load_raw_ticks: data.rawCurrentLoad,
-      }))
-      .then(data => {
-        const cpuLoadMetrics = [];
-        for (const metric in data) {
-          cpuLoadMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'CPU',
-            time,
-          });
-        }
-        return cpuLoadMetrics;
-      })
-      .catch(err => {
+  si.cpu()
+    .then(data => {
+      const siMethodName = 'cpu';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'CPU',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
         throw err;
-      })
-  );
+      }
+    });
 
-  promises.push(
-    si
-      .mem()
-      .then(data => ({
-        totalmemory_in_bytes: data.total,
-        freememory_in_bytes: data.free,
-        usedmemory_in_bytes: data.used,
-        activememory_in_bytes: data.active,
-        buffers_plus_cache_in_bytes: data.buffcache,
-        available_memory: data.available,
-      }))
-      .then(data => {
-        const memMetrics = [];
-        for (const metric in data) {
-          memMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'Memory',
-            time,
-          });
-        }
-        return memMetrics;
-      })
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
+  si.cpuCurrentSpeed()
+    .then(data => {
+      const siMethodName = 'cpuCurrentSpeed';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'CPU',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
 
-  promises.push(
-    si
-      .processes()
-      .then(data => ({
-        totalprocesses: data.all,
-        blockedprocesses: data.blocked,
-        runningprocesses: data.running,
-        sleepingprocesses: data.sleeping,
-      }))
-      .then(data => {
-        const processMetrics = [];
-        for (const metric in data) {
-          processMetrics.push({
-            metric,
-            value: data[metric],
-            category: 'Processes',
-            time,
-          });
-        }
-        return processMetrics;
-      })
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
+  si.cpuTemperature()
+    .then(data => {
+      const siMethodName = 'cpuTemperature';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'CPU',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
 
-  promises.push(
-    si
-      .inetLatency()
-      .then(data => ({
+  si.currentLoad()
+    .then(data => {
+      const siMethodName = 'currentLoad';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'CPU',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
+
+  si.mem()
+    .then(data => {
+      const siMethodName = 'mem';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'Memory',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
+
+  si.processes()
+    .then(data => {
+      const siMethodName = 'processes';
+      for (let metricName in collectedMetrics[siMethodName]) {
+        healthDataCollection.push({
+          metric: metricName,
+          value: data[collectedMetrics[siMethodName][metricName]],
+          category: 'Processes',
+          time,
+        });
+      }
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
+
+  si.inetLatency()
+    .then(data => {
+      const siMethodName = 'inetLatency';
+      healthDataCollection.push({
         metric: 'latency',
         value: data,
-        category: 'Latency',
+        category: 'Memory',
         time,
-      }))
-      .catch(err => {
-        if (err) {
-          throw err;
-        }
-      })
-  );
+      });
+    })
+    .catch(err => {
+      if (err) {
+        throw err;
+      }
+    });
+
   // Return a promise that resolves to an array of all of the data points unnested
-  return Promise.all(promises)
-    .then(array => array.flat())
-    // Remove any empty strings, NaN, or "NaN" from values prevent database errors
-    .then((array) => array.filter((a) => {
-      if (isNaN(a.value) || a.value === 'NaN' || a.value === '') return false;
-      else return true;
-    }));
+  return (
+    Promise.all(healthDataCollection)
+      // Remove any empty strings, NaN, or "NaN" from values prevent database errors
+      .then(array =>
+        array.filter(metric => {
+          if (
+            isNaN(metric.value) ||
+            metric.value === 'NaN' ||
+            metric.value === '' ||
+            metric.value === null
+          )
+            return false;
+          else return true;
+        })
+      )
+  );
 };
 
 module.exports = healthHelpers;
